feat(TS-003): added unit tests for products

+++ 2 Domain/ProductDomain.cs
@@ -13,13 +13,17 @@ public class ProductDomain : IProductDomain
         _productData = productData;
     }
 
-    public async Task<bool> SaveAsync(Product data)
+    public async Task<int> SaveAsync(Product data)
     {
-        //add business rules
-        var product = await _productData.GetByNameAsync(data.Name);
-        if (product!= null)
+        
+        var existingProduct = await _productData.GetByNameAsync(data.Name);
+        if (existingProduct!= null)
         {
             throw new Exception("Name already exists.");
+        }
+        if(data.CategoryId == 0)
+        {
+            throw new Exception("Category is required.");
         }
         
         return await _productData.SaveAsync(data);
@@ -32,6 +36,12 @@ public class ProductDomain : IProductDomain
 
     public async Task<bool> DeleteAsync(int id)
     {
+        var existingProduct = _productData.GetById(id);
+        if(existingProduct == null)
+        {
+            throw new Exception("Product not found.");
+        }
+        
         return await _productData.DeleteAsync(id);
     }
